browser/components/backup/tests/chrome/test_enable_backup_encryption.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Tests for the enable-backup-encryption component</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <script
  src="chrome://browser/content/backup/enable-backup-encryption.mjs"
  type="module"
></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script>

    const { BrowserTestUtils } = ChromeUtils.importESModule(
      "resource://testing-common/BrowserTestUtils.sys.mjs"
    );
    const MOCK_PASSWORD = "mckP@ss3x2 fake_password";

    function createMockPassInputEventPromise(inputEl, mockPassword) {
      let promise = new Promise(resolve => {
        inputEl.addEventListener("input", () => resolve(), {
          once: true,
        });
      });
      inputEl.focus();
      inputEl.value = mockPassword;
      inputEl.dispatchEvent(new Event("input"));
      return promise;
    }

    /**
     * Tests that adding a enable-backup-encryption element to the DOM causes it to
     * fire a BackupUI:InitWidget event.
     */
    add_task(async function test_initWidget() {
      let enableBackupEncryption = document.createElement("enable-backup-encryption");
      let content = document.getElementById("content");

      let sawInitWidget = BrowserTestUtils.waitForEvent(content, "BackupUI:InitWidget");
      content.appendChild(enableBackupEncryption);
      await sawInitWidget;
      ok(true, "Saw BackupUI:InitWidget");

      enableBackupEncryption.remove();
    });

    /**
     * Tests that pressing the confirm button for the set-password type dialog will dispatch the expected events.
     */
    add_task(async function test_type_set_password_confirm() {
      let enableBackupEncryption = document.getElementById("test-enable-backup-encryption");

      let confirmButton = enableBackupEncryption.confirmButtonEl;
      ok(confirmButton, "Confirm button should be found");
      ok(confirmButton.disabled, "Confirm button should now be disabled since there are passwords expanded options");

      let newPasswordInput = enableBackupEncryption.inputNewPasswordEl;
      let repeatPasswordInput = enableBackupEncryption.inputRepeatPasswordEl;
      ok(newPasswordInput, "New password input should be found");
      ok(repeatPasswordInput, "Repeat password input should be found");

      // Pretend we're entering a password in the new password field
      let newPassPromise = createMockPassInputEventPromise(newPasswordInput, MOCK_PASSWORD)
      await newPassPromise;
      confirmButton = enableBackupEncryption.confirmButtonEl;
      ok(confirmButton.disabled, "Confirm button should still be disabled with no repeat password");

      // Pretend we're entering a password in the repeat field
      // Passwords do not match
      const tempPassword = `${MOCK_PASSWORD}-notMatch`;
      let promiseNotMatchPass = createMockPassInputEventPromise(repeatPasswordInput, tempPassword);
      await promiseNotMatchPass;
      confirmButton = enableBackupEncryption.confirmButtonEl;
      ok(confirmButton.disabled, "Confirm button should still be disabled with no matching passwords");

      // Passwords match
      let promiseMatchPass = createMockPassInputEventPromise(repeatPasswordInput, MOCK_PASSWORD);
      await promiseMatchPass;

      confirmButton = enableBackupEncryption.confirmButtonEl;

      let confirmButtonPromise = BrowserTestUtils.waitForMutationCondition(
        confirmButton,
        { attributes: true },
        () => !confirmButton.disabled
      );

      await confirmButtonPromise;
      ok(!confirmButton.disabled, "Confirm button should no longer be disabled");

      let content = document.getElementById("content");
      let encryptionPromise = BrowserTestUtils.waitForEvent(content, "enableEncryption");

      confirmButton.click()

      await encryptionPromise;
      ok(true, "Detected event after selecting the confirm button");
    })

    /**
     * Tests that pressing the confirm button for the set-password type dialog will dispatch the expected events.
     */
     add_task(async function test_type_change_password_confirm() {
      let enableBackupEncryption = document.getElementById("test-enable-backup-encryption");
      enableBackupEncryption.type = "change-password";
      await enableBackupEncryption.updateComplete;

      let confirmButton = enableBackupEncryption.confirmButtonEl;
      ok(confirmButton, "Confirm button should be found");
      ok(confirmButton.disabled, "Confirm button should now be disabled since there are passwords expanded options");

      let newPasswordInput = enableBackupEncryption.inputNewPasswordEl;
      let repeatPasswordInput = enableBackupEncryption.inputRepeatPasswordEl;
      ok(newPasswordInput, "New password input should be found");
      ok(repeatPasswordInput, "Repeat password input should be found");

      // Pretend we're entering a password in the new password field
      let newPassPromise = createMockPassInputEventPromise(newPasswordInput, MOCK_PASSWORD);
      await newPassPromise;
      confirmButton = enableBackupEncryption.confirmButtonEl;
      ok(confirmButton.disabled, "Confirm button should still be disabled with no repeat password");

      // Pretend we're entering a password in the repeat field
      // Passwords do not match
      const tempPassword = `${MOCK_PASSWORD}-notMatch`;
      let promiseNotMatchPass = createMockPassInputEventPromise(repeatPasswordInput, tempPassword);
      await promiseNotMatchPass;
      confirmButton = enableBackupEncryption.confirmButtonEl;
      ok(confirmButton.disabled, "Confirm button should still be disabled with no matching passwords");

      // Passwords match
      let promiseMatchPass = createMockPassInputEventPromise(repeatPasswordInput, MOCK_PASSWORD);
      await promiseMatchPass;

      confirmButton = enableBackupEncryption.confirmButtonEl;

      let confirmButtonPromise = BrowserTestUtils.waitForMutationCondition(
        confirmButton,
        { attributes: true },
        () => !confirmButton.disabled
      );

      await confirmButtonPromise;
      ok(!confirmButton.disabled, "Confirm button should no longer be disabled");

      let content = document.getElementById("content");
      let encryptionPromise = BrowserTestUtils.waitForEvent(content, "rerunEncryption");

      confirmButton.click()

      await encryptionPromise;
      ok(true, "Detected event after selecting the confirm button");
    })

    /**
     * Tests that changing the dialog type to "change-password" will hide the text description.
     */
     add_task(async function test_type_change_password_description() {
      let enableBackupEncryption = document.getElementById("test-enable-backup-encryption");
      enableBackupEncryption.type = "change-password";
      await enableBackupEncryption.updateComplete;

      let textDescription = enableBackupEncryption.textDescriptionEl;
      ok(!textDescription, "The dialog text description should not be visible if type is change-password");
    })

    /**
     * Tests that pressing the cancel button will dispatch the expected events.
     */
    add_task(async function test_cancel() {
      let enableBackupEncryption = document.getElementById("test-enable-backup-encryption");
      let cancelButton = enableBackupEncryption.cancelButtonEl;

      ok(cancelButton, "Cancel button should be found");

      let content = document.getElementById("content");
      let promise = BrowserTestUtils.waitForEvent(content, "dialogCancel");

      cancelButton.click()

      await promise;
      ok(true, "Detected event after selecting the cancel button");
    })
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
  <enable-backup-encryption id="test-enable-backup-encryption"></enable-backup-encryption>
</div>
<pre id="test"></pre>
</body>
</html>
